% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ipw_viz_point}
\alias{ipw_viz_point}
\title{Creating scatterplots to determine model specification for propensity scoring (linear or non-linear) - the blue smooth reflects observed data trends, and the red smooth shows the predictions from a logistic regression model with a specific model formula}
\usage{
ipw_viz_point(data, x, y)
}
\arguments{
\item{data}{the data used for the propensity scoring}

\item{x}{the x axis for the scatterplot (the variable whose specification you're exploring)}

\item{y}{the y axis for the scatterplot (the outcome variable you're interested in)}
}
\description{
Creating scatterplots to determine model specification for propensity scoring (linear or non-linear) - the blue smooth reflects observed data trends, and the red smooth shows the predictions from a logistic regression model with a specific model formula
}
\examples{
ipw_viz_point(tenure, revenue, gncs)
ipw_viz_point(tenure, full_ratio, gncs)
}
