% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{odds_to_prob}
\alias{odds_to_prob}
\title{Changing odds to a probability}
\usage{
odds_to_prob(odds)
}
\arguments{
\item{odds}{A number in odds form}
}
\description{
Changing odds to a probability
}
\examples{
prob = odds_to_prob(log_odds_A)
odds_to_prob(-0.6)
}
