% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_z_in_chain}
\alias{set_z_in_chain}
\title{Simulates Z values that are dependent on Y values in a chain. This function is used within the set_y_and_z_in_chain(n, y_numeric, z_numeric, X) function.}
\usage{
set_z_in_chain(n, z_numeric, Y)
}
\arguments{
\item{n}{The number of observations of Z for a single simulation}

\item{z_numeric}{TRUE for numeric Z and FALSE for binary Z}

\item{Y}{The values of Y in the big function to generate Z values that are dependent on Y values}
}
\value{
Simulated Z values
}
\description{
Simulates Z values that are dependent on Y values in a chain. This function is used within the set_y_and_z_in_chain(n, y_numeric, z_numeric, X) function.
}
